requirements.txt
absl-py==1.4.0
appdirs==1.4.4
astunparse==1.6.3
audioread==3.0.0
blinker==1.6.2
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
decorator==5.1.1
Flask==2.3.1
flatbuffers==23.3.3
gast==0.4.0
google-auth==2.17.3
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.54.0
gunicorn==20.1.0
h5py==3.8.0
idna==3.4
itsdangerous==2.1.2
jax==0.4.8
Jinja2==3.1.2
joblib==1.2.0
keras==2.12.0
lazy_loader==0.2
libclang==16.0.0
librosa==0.10.0.post2
llvmlite==0.39.1
Markdown==3.4.3
MarkupSafe==2.1.2
ml-dtypes==0.1.0
msgpack==1.0.5
numba==0.56.4
numpy==1.23.5
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.1
pooch==1.6.0
protobuf==4.22.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
requests==2.29.0
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.2.2
scipy==1.10.1
six==1.16.0
soundfile==0.12.1
soxr==0.3.5
tensorboard==2.12.2
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tensorflow-estimator==2.12.0
tensorflow-io-gcs-filesystem==0.32.0
termcolor==2.3.0
threadpoolctl==3.1.0
typing_extensions==4.5.0
urllib3==1.26.15
Werkzeug==2.3.1
wrapt==1.14.1
